medical-ai-platform/ml-service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
Pillow>=10.1.0
numpy>=1.26.0
opencv-python-headless>=4.8.0
python-multipart>=0.0.6
